+++ examples/Console/Program.cs
@@ -32,7 +32,7 @@ namespace Examples.Console
         /// dotnet run -p Examples.Console.csproj zipkin -u http://localhost:9411/api/v2/spans
         /// dotnet run -p Examples.Console.csproj jaeger -h localhost -p 6831
         /// dotnet run -p Examples.Console.csproj prometheus -i 15 -p 9184 -d 2
-        /// dotnet run -p Examples.Console.csproj otlp -e "localhost:55680"
+        /// dotnet run -p Examples.Console.csproj otlp -e "localhost:4317"
         /// dotnet run -p Examples.Console.csproj zpages
         ///
         /// The above must be run from the project root folder
@@ -119,7 +119,7 @@ namespace Examples.Console
     [Verb("otlp", HelpText = "Specify the options required to test OpenTelemetry Protocol (OTLP)")]
     internal class OtlpOptions
     {
-        [Option('e', "endpoint", HelpText = "Target to which the exporter is going to send traces or metrics", Default = "http://localhost:55680")]
+        [Option('e', "endpoint", HelpText = "Target to which the exporter is going to send traces or metrics", Default = "http://localhost:4317")]
         public string Endpoint { get; set; }
     }