+++ src/OpenTelemetry.Exporter.Jaeger/CHANGELOG.md
@@ -5,8 +5,8 @@
 * Moved `JaegerExporter` and `JaegerExporterOptions` classes to
   `OpenTelemetry.Exporter` namespace.
   ([#1770](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1770))
-* Default `service.name` complies with OTel spec. Grabs default from
-  `ParentProvider` using GetDefaultResource() APIfrom SDK if not found.
+* Default ServiceName, if not found in Resource is obtained from SDK
+  using GetDefaultResource().
   [#1768](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1768)
 
 ## 1.0.0-rc2